Move 'Cumulatives' to Provider workspace
