+++ src/Server.UI/Pages/Workspaces/Provider/Components/CumulativeDetailsComponent.razor.css
