+++ src/Server.UI/Pages/Workspaces/Provider/Components/ProviderPageHeader.razor
@@ -29,6 +29,7 @@
             "case-workload" => [home, ProviderLinks.CaseWorkload.AsMudBreadcrumb()],
             "case-management" => [home, ProviderLinks.CaseManagement.AsMudBreadcrumb()],
             "performance" => [home, ProviderLinks.Performance.AsMudBreadcrumb()],
+            "cumulatives" => [home, ProviderLinks.Cumulatives.AsMudBreadcrumb()],
             _ => throw new Exception()
         };
     }
