+++ src/Server.UI/Pages/Workspaces/Provider/Pages/Cumulatives.razor
@@ -1,16 +1,12 @@
 ﻿@using Cfo.Cats.Application.SecurityConstants
-@using Cfo.Cats.Server.UI.Pages.Workspaces.Performance.Components
-
+@using Cfo.Cats.Server.UI.Pages.Workspaces.Provider.Components
 
 @inherits CatsComponentBase
 @attribute [Authorize(Roles = $"{RoleNames.SystemSupport}, {RoleNames.Finance}")]
 
-@page "/pages/analytics/cumulatives"
+@page "/pages/workspace/provider/cumulatives"
 
-
-<MudText Typo="Typo.h5">
-    Cumulatives
-</MudText>
+<ProviderPageHeader />
 
 @if (_noAccessToContracts)
 {
