+++ src/Server.UI/Pages/Workspaces/Provider/Pages/Cumulatives.razor.cs
@@ -4,7 +4,7 @@ using Cfo.Cats.Application.Features.Contracts.DTOs;
 using Cfo.Cats.Application.Features.ManagementInformation.Commands;
 using Cfo.Cats.Infrastructure.Constants;
 
-namespace Cfo.Cats.Server.UI.Pages.Analytics;
+namespace Cfo.Cats.Server.UI.Pages.Workspaces.Provider.Pages;
 
 public partial class Cumulatives
 {
